Changed default data type to np.float64

+++ src/mfnet/tensor.py
@@ -1,7 +1,7 @@
 from functools import partial
 
-from numpy import array, dtype, float32, ndarray
+from numpy import array, dtype, float64, ndarray
 
-type Tensor = ndarray[tuple[int, ...], dtype[float32]]
+type Tensor = ndarray[tuple[int, ...], dtype[float64]]
 
-tensor = partial(array, dtype=float32)
+tensor = partial(array, dtype=float64)
